refactor: add generic accumulator class to satisfy adl in concept ( PROOF-589 )
