+++ sxt/algorithm/base/accumulator.h
@@ -18,18 +18,16 @@
 
 #include <concepts>
 
-namespace sxt::bascrv {
+#include "sxt/base/macro/cuda_callable.h"
+
+namespace sxt::algb {
 //--------------------------------------------------------------------------------------------------
-// element
+// accumulator
 //--------------------------------------------------------------------------------------------------
-template <class T>
-concept element = requires(T& res, const T& e) {
-  double_element(res, e);
-  add(res, e, e);
-  neg(res, e);
-  add_inplace(res, res);
-  { T::identity() } noexcept -> std::same_as<T>;
-  mark(res);
-  { is_marked(e) } noexcept -> std::same_as<bool>;
+template <class T> class accumulator {
+public:
+  using value_type = T;
+
+  CUDA_CALLABLE static void accumulate_inplace(T& res, T& e) noexcept { add_inplace(res, e); }
 };
-} // namespace sxt::bascrv
+} // namespace sxt::algb
