+++ sxt/multiexp/curve/multiproduct.h
@@ -16,26 +16,21 @@
  */
 #pragma once
 
+#include "sxt/algorithm/base/accumulator.h"
 #include "sxt/algorithm/base/gather_mapper.h"
 #include "sxt/base/container/span.h"
 #include "sxt/base/curve/element.h"
 #include "sxt/base/macro/cuda_callable.h"
-#include "sxt/curve21/operation/accumulator.h"
 #include "sxt/execution/async/future.h"
 #include "sxt/execution/async/future_fwd.h"
 #include "sxt/multiexp/multiproduct_gpu/multiproduct.h"
-
-namespace sxt::basdv {
-class stream;
-}
 
 namespace sxt::mtxcrv {
 //--------------------------------------------------------------------------------------------------
 // signed_mapper
 //--------------------------------------------------------------------------------------------------
 namespace {
-template <bascrv::element Element>
-class signed_mapper {
+template <bascrv::element Element> class signed_mapper {
 public:
   using value_type = Element;
 
@@ -67,18 +62,16 @@ private:
 // async_compute_multiproduct
 //--------------------------------------------------------------------------------------------------
 template <bascrv::element Element>
-inline xena::future<> async_compute_multiproduct(basct::span<Element> products,
-                                                 const basdv::stream& stream,
-                                                 basct::cspan<Element> generators,
-                                                 basct::cspan<unsigned> indexes,
-                                                 basct::cspan<unsigned> product_sizes,
-                                                 bool is_signed) noexcept {
+inline xena::future<>
+async_compute_multiproduct(basct::span<Element> products, const basdv::stream& stream,
+                           basct::cspan<Element> generators, basct::cspan<unsigned> indexes,
+                           basct::cspan<unsigned> product_sizes, bool is_signed) noexcept {
   if (!is_signed) {
     using Mapper = algb::gather_mapper<Element>;
-    return mtxmpg::compute_multiproduct<c21o::accumulator, Mapper>(products, stream, generators,
-                                                                   indexes, product_sizes);
+    return mtxmpg::compute_multiproduct<algb::accumulator<Element>, Mapper>(
+        products, stream, generators, indexes, product_sizes);
   }
-  return mtxmpg::compute_multiproduct<c21o::accumulator, signed_mapper<Element>>(
+  return mtxmpg::compute_multiproduct<algb::accumulator<Element>, signed_mapper<Element>>(
       products, stream, generators, indexes, product_sizes);
 }
 } // namespace sxt::mtxcrv
